requirements.txt
discord.py==2.2.2
redis==4.5.3
